pyproject.toml
[project]
name = "lms-connect-system"
version = "0.1.0"
description = ""
authors = [
    {name = "GuyMonagan",email = "werevalkury@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "django-filter (>=25.2,<26.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "psycopg[binary] (>=3.3.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

package-mode = false